test(e2e): split phase-3 journey into per-endpoint files
Previous file phase-3-headless-mutations.journey.ts grouped three
unrelated endpoints under a Phase-named header — broke per-endpoint
convention (tasks-claim.journey.ts, workflow-namespacing.journey.ts,
etc.) and 'Phase 3' in the filename rots over time. Split:

- tasks-complete.journey.ts: verdict variant + Phase 3 audit chain
  assertions (task.completed + process.resumed_after_task). Sibling to
  tasks-claim.journey.ts. previous-run-outputs.journey.ts already covers
  the params variant; this file covers verdict.

- cron-heartbeat.journey.ts: 3 tests — auth 401, response shape,
  'skipped triggers do not re-fire' invariant (back-to-back heartbeats).

- previous-run-outputs.journey.ts: add explicit entity-echo regression
  asserts (.instanceId/.status absent at top level) where the journey
  already starts a run. Catches accidental revert of ADR-0005 §5 without
  a separate file.
