--- packages/platform-ui/e2e/api/phase-3-headless-mutations.journey.ts
@@ -1,288 +1,0 @@
-import { test, expect } from '../helpers/test-fixtures';
-import { TEST_ORG_HANDLE } from '../helpers/constants';
-
-/**
- * Phase 3 of the headless-platform-API migration. Three E2E API journeys
- * exercise the new contracts against a running platform via apiKey:
- *
- *   1. `tasks/complete` discriminated-union body — full audit assertions.
- *   2. `POST /api/processes` entity-echo response shape (regression for
- *      the pre-Phase-3 `{ instanceId, status }` → `{ run }` break).
- *   3. `cron/heartbeat` auth + response shape + "skipped triggers do NOT
- *      audit" invariant.
- */
-
-const API_KEY = process.env.PLATFORM_API_KEY ?? 'test-api-key';
-
-async function pollUntil<T>(
-  fn: () => Promise<T | null>,
-  {
-    timeoutMs = 10_000,
-    intervalMs = 200,
-    description = 'condition',
-  }: { timeoutMs?: number; intervalMs?: number; description?: string } = {},
-): Promise<T> {
-  const deadline = Date.now() + timeoutMs;
-  let last: T | null = null;
-  while (Date.now() < deadline) {
-    last = await fn();
-    if (last !== null) return last;
-    await new Promise((r) => setTimeout(r, intervalMs));
-  }
-  throw new Error(`Timed out waiting for ${description} (${timeoutMs}ms)`);
-}
-
-test.describe('Phase 3 — headless mutations API E2E', () => {
-  test('tasks/complete verdict variant: full flow + audit chain', async ({ request }) => {
-    const wdName = `e2e-p3-verdict-${Date.now()}`;
-
-    // Two-step WD: human review step → terminal. After "approve", advanceStep
-    // takes the run to terminal → instance.status === 'completed'.
-    const wd = {
-      name: wdName,
-      title: 'Phase 3 verdict E2E',
-      steps: [
-        {
-          id: 'review',
-          name: 'Review',
-          type: 'review',
-          executor: 'human',
-          allowedRoles: ['operator'],
-        },
-        { id: 'done', name: 'Done', type: 'terminal', executor: 'human' },
-      ],
-      transitions: [{ from: 'review', to: 'done' }],
-      triggers: [{ type: 'manual', name: 'Start' }],
-    };
-
-    const createWdRes = await request.post(
-      `/api/workflow-definitions?namespace=${TEST_ORG_HANDLE}`,
-      {
-        headers: { 'X-Api-Key': API_KEY, 'Content-Type': 'application/json' },
-        data: wd,
-      },
-    );
-    expect(createWdRes.status(), await createWdRes.text()).toBe(201);
-
-    // POST /api/processes returns the new `{ run }` entity-echo shape (201).
-    const triggerRes = await request.post('/api/processes', {
-      headers: { 'X-Api-Key': API_KEY, 'Content-Type': 'application/json' },
-      data: {
-        namespace: TEST_ORG_HANDLE,
-        definitionName: wdName,
-        triggeredBy: 'e2e-test',
-        triggerName: 'Start',
-      },
-    });
-    expect(triggerRes.status(), await triggerRes.text()).toBe(201);
-    const startBody = (await triggerRes.json()) as {
-      run: { id: string; status: string };
-    };
-    expect(startBody.run).toBeDefined();
-    expect(typeof startBody.run.id).toBe('string');
-    const instanceId = startBody.run.id;
-
-    // Wait for the review task — engine.advanceStep creates it after start.
-    const task = await pollUntil(
-      async () => {
-        const res = await request.get(`/api/tasks?instanceId=${instanceId}`, {
-          headers: { 'X-Api-Key': API_KEY },
-        });
-        if (res.status() !== 200) return null;
-        const body = (await res.json()) as {
-          tasks: Array<{ id: string; stepId: string; status: string }>;
-        };
-        return body.tasks.find((t) => t.stepId === 'review') ?? null;
-      },
-      { description: `review task for ${instanceId}` },
-    );
-
-    // Complete via the new discriminated-union body.
-    const completeRes = await request.post(`/api/tasks/${task.id}/complete`, {
-      headers: { 'X-Api-Key': API_KEY, 'Content-Type': 'application/json' },
-      data: { kind: 'verdict', verdict: 'approve', comment: 'lgtm' },
-    });
-    expect(completeRes.status(), await completeRes.text()).toBe(200);
-    const completeBody = (await completeRes.json()) as {
-      task: { id: string; status: string };
-      run: { id: string; status: string };
-    };
-    expect(completeBody.task.status).toBe('completed');
-    expect(completeBody.run.id).toBe(instanceId);
-
-    // Run advances to terminal and completes.
-    await pollUntil(
-      async () => {
-        const res = await request.get(`/api/processes/${instanceId}`, {
-          headers: { 'X-Api-Key': API_KEY },
-        });
-        if (res.status() !== 200) return null;
-        const body = (await res.json()) as { status: string };
-        return body.status === 'completed' ? body : null;
-      },
-      { description: `instance ${instanceId} to complete` },
-    );
-
-    // Audit chain assertions: every expected row from Phase 3 must be present.
-    const auditRes = await request.get(`/api/processes/${instanceId}/audit`, {
-      headers: { 'X-Api-Key': API_KEY },
-    });
-    expect(auditRes.status()).toBe(200);
-    const { events } = (await auditRes.json()) as {
-      events: Array<{ action: string; basis?: string }>;
-    };
-    const actions = events.map((e) => e.action);
-    // Engine-emitted lifecycle:
-    expect(actions).toContain('instance.created');
-    expect(actions).toContain('instance.started');
-    expect(actions).toContain('task.created');
-    // Phase 3 handler-resident emits (ADR-0005 §7):
-    expect(actions).toContain('task.completed');
-    expect(actions).toContain('process.resumed_after_task');
-  });
-
-  test('POST /api/processes entity echo: response is `{ run }`, not `{ instanceId, status }`', async ({
-    request,
-  }) => {
-    const wdName = `e2e-p3-entity-echo-${Date.now()}`;
-    const wd = {
-      name: wdName,
-      title: 'Phase 3 entity echo',
-      steps: [
-        { id: 'wait', name: 'Wait', type: 'creation', executor: 'human' },
-        { id: 'done', name: 'Done', type: 'terminal', executor: 'human' },
-      ],
-      transitions: [{ from: 'wait', to: 'done' }],
-      triggers: [{ type: 'manual', name: 'Start' }],
-    };
-    const createWdRes = await request.post(
-      `/api/workflow-definitions?namespace=${TEST_ORG_HANDLE}`,
-      {
-        headers: { 'X-Api-Key': API_KEY, 'Content-Type': 'application/json' },
-        data: wd,
-      },
-    );
-    expect(createWdRes.status()).toBe(201);
-
-    const triggerRes = await request.post('/api/processes', {
-      headers: { 'X-Api-Key': API_KEY, 'Content-Type': 'application/json' },
-      data: {
-        namespace: TEST_ORG_HANDLE,
-        definitionName: wdName,
-        triggeredBy: 'e2e-test',
-        triggerName: 'Start',
-      },
-    });
-    expect(triggerRes.status()).toBe(201);
-    const body = (await triggerRes.json()) as Record<string, unknown>;
-
-    // Old shape (Phase ≤2) returned `{ instanceId, status }` at the top
-    // level. Phase 3 ADR-0005 §5 entity echo returns `{ run: ProcessInstance }`.
-    // Both invariants checked: new key present + old keys absent.
-    expect(body.run).toBeDefined();
-    expect(body.instanceId).toBeUndefined();
-    expect(body.status).toBeUndefined();
-
-    const run = body.run as Record<string, unknown>;
-    expect(typeof run.id).toBe('string');
-    expect(typeof run.status).toBe('string');
-    expect(typeof run.namespace).toBe('string');
-    expect(run.namespace).toBe(TEST_ORG_HANDLE);
-    expect(run.definitionName).toBe(wdName);
-  });
-
-  test('cron/heartbeat: auth required + skipped triggers do NOT audit', async ({
-    request,
-  }) => {
-    // (a) Missing API key → 401 from middleware.
-    const noAuthRes = await request.post('/api/cron/heartbeat');
-    expect(noAuthRes.status()).toBe(401);
-
-    // (b) With API key (system actor) → 200 with the contracted shape.
-    const okRes = await request.post('/api/cron/heartbeat', {
-      headers: { 'X-Api-Key': API_KEY },
-    });
-    expect(okRes.status(), await okRes.text()).toBe(200);
-    const body = (await okRes.json()) as {
-      triggered: Array<{
-        definitionName: string;
-        triggerName: string;
-        instanceId: string;
-      }>;
-      skipped: Array<{
-        definitionName: string;
-        triggerName: string;
-        reason: string;
-      }>;
-    };
-    expect(Array.isArray(body.triggered)).toBe(true);
-    expect(Array.isArray(body.skipped)).toBe(true);
-
-    // (c) "Skipped triggers do not audit" invariant — load-bearing per
-    // ADR-0005 §7 ("emit only on state change"). Take a global audit count
-    // before + after a heartbeat that only skips; counts must match for the
-    // skipped rows. Strategy: register a cron WD with an obviously-due
-    // trigger, fire heartbeat (audits one row), fire again (now state is
-    // recent → "Not due" → must NOT audit). Compare the second-call audit
-    // delta against the trigger instance's own audit chain.
-    const wdName = `e2e-p3-cron-${Date.now()}`;
-    const cronWd = {
-      name: wdName,
-      title: 'Phase 3 cron skip',
-      steps: [
-        { id: 'noop', name: 'Noop', type: 'creation', executor: 'human' },
-        { id: 'done', name: 'Done', type: 'terminal', executor: 'human' },
-      ],
-      transitions: [{ from: 'noop', to: 'done' }],
-      triggers: [
-        { type: 'cron', name: 'every-15m', schedule: '*/15 * * * *' },
-      ],
-    };
-    const createWdRes = await request.post(
-      `/api/workflow-definitions?namespace=${TEST_ORG_HANDLE}`,
-      {
-        headers: { 'X-Api-Key': API_KEY, 'Content-Type': 'application/json' },
-        data: cronWd,
-      },
-    );
-    expect(createWdRes.status()).toBe(201);
-
-    // First heartbeat — schedule is freshly created so isDue() returns true
-    // (no prior state → uses def.createdAt as the floor → 15m has passed
-    // since "the Unix epoch" of this WD, which is right now, so this
-    // assertion is brittle if isDue() compares to def.createdAt with a
-    // sub-minute granularity). Skip the "must fire" assertion; only assert
-    // that the second call audits nothing new.
-    await request.post('/api/cron/heartbeat', {
-      headers: { 'X-Api-Key': API_KEY },
-    });
-
-    // Re-fire heartbeat immediately — schedule is */15min and state was
-    // just persisted, so this must skip with "Not due" for our trigger.
-    const second = await request.post('/api/cron/heartbeat', {
-      headers: { 'X-Api-Key': API_KEY },
-    });
-    expect(second.status()).toBe(200);
-    const secondBody = (await second.json()) as {
-      triggered: Array<{ definitionName: string; triggerName: string }>;
-      skipped: Array<{
-        definitionName: string;
-        triggerName: string;
-        reason: string;
-      }>;
-    };
-    const ourSkip = secondBody.skipped.find(
-      (s) => s.definitionName === wdName && s.triggerName === 'every-15m',
-    );
-    expect(ourSkip).toBeDefined();
-    expect(ourSkip?.reason).toBe('Not due');
-
-    // If a first-call fire created an instance, the skip on the second call
-    // must NOT have written a new audit row to that instance — verifies the
-    // "emit only on state change" invariant for cron skips.
-    const ourTriggered = secondBody.triggered.find(
-      (t) => t.definitionName === wdName && t.triggerName === 'every-15m',
-    );
-    expect(ourTriggered).toBeUndefined();
-  });
-});
